+++ examples/adverbial.R
@@ -1,0 +1,2 @@
+nom_adv(1:4)
+nom_adv(1:4, thrice = TRUE)