+++ NEWS.md
@@ -7,9 +7,16 @@
   This is useful when you want to print small numbers in words but larger
   numbers numerically.
   Refactorization means this argument can be used by all functions.
+  - `options("nombre.max_n")` sets a default value.
+  
+* `adverbial()` and `nom_adv()` generate adverbials, e.g. "once", "twice",
+  "three times".
+  - `thrice = TRUE` or `options("nombre.thrice" = TRUE)` convert `3` to "thrice"
+    instead of "three times".
   
 * All functions now pass `...` to lower level functions.
-  Passing follows the sequence `denominator()` -> `ordinal()` -> `cardinal()`.
+  `ordinal()` and `adverbial()` pass to `cardinal()`.
+  `denominator()` passes to `ordinal()`.
   This allows `cardinal()`'s `max_n` and `negative` to be used by all functions
   and `ordinal()`'s `cardinal` to be used by `denominator()`.